requirements.txt
pycryptodome==3.9.7
langdetect==1.14.0
boto3==1.13.9
googletrans==2.4.0
google-cloud-translate==2.0.1
requests==2.23.0
